+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- Introduce `-a` and `-f` options for aging queries to `vast start`.
+
 ### Changed
 - Fix evaluation of predicates with negations and type or key extractors.